Move `EthereumTransaction` to support_ethereum
