+++ task/task-darwinia-ethereum/src/service/ethereum/ethereum_logs_handler.rs
@@ -6,11 +6,11 @@ use web3::types::{Log, H160, H256};
 
 use bridge_traits::bridge::task::BridgeSand;
 use component_darwinia_subxt::darwinia::client::Darwinia;
+use support_ethereum::transaction::{EthereumTransaction, EthereumTransactionHash};
 use support_tracker::Tracker;
 use support_tracker_evm_log::{EvmClient, LogsHandler, Result};
 
 use crate::message::{ToRedeemMessage, ToRelayMessage};
-use crate::service::{EthereumTransaction, EthereumTransactionHash};
 use crate::task::DarwiniaEthereumTask;
 
 const MAX_WAITED_REDEEM_COUNT: usize = 1024;
