+++ task/task-pangolin-ropsten/src/service/ropsten/ropsten_logs_handler.rs
@@ -1,15 +1,16 @@
+use std::collections::{HashMap, HashSet};
+
 use lifeline::Sender;
 use postage::broadcast;
-use std::collections::{HashMap, HashSet};
 use web3::types::{Log, H160, H256};
 
 use bridge_traits::bridge::task::BridgeSand;
 use component_pangolin_subxt::darwinia::client::Darwinia;
+use support_ethereum::transaction::{EthereumTransaction, EthereumTransactionHash};
 use support_tracker::Tracker;
 use support_tracker_evm_log::{EvmClient, LogsHandler, Result};
 
 use crate::message::{ToRedeemMessage, ToRelayMessage};
-use crate::service::{EthereumTransaction, EthereumTransactionHash};
 use crate::task::PangolinRopstenTask;
 
 const MAX_WAITED_REDEEM_COUNT: usize = 1024;
